changed shortcut urls to use new trunk url, they should grab this from .svn or something though

+++ Core/SoarKernel/src/exploration.cpp
@@ -66,7 +66,7 @@ const long exploration_convert_policy( const char *policy_name )
 	if ( !strcmp( policy_name, "softmax" ) )
 		return USER_SELECT_SOFTMAX;
 
-	return long(NULL);
+	return NULL;
 }
 
 const char *exploration_convert_policy( const long policy )
